production/streamlit-server/requirements.txt
streamlit==0.86.0
requests==2.26.0
python-decouple==3.4
